Login, delete, Read are fully functional
