+++ API/domain.json
@@ -14,12 +14,6 @@
       "sideboardNotes": "miracles sb: out: -4 mother of runes, 2 swords to plowshares in: 2 Gideon, 2 councils Judgement, 1 leonin-relic, 1 pithing needle"
     },
     {
-      "id": 2,
-      "name": "Miracles",
-      "user_id": 1,
-      "sideboardNotes": "miracles sb: out: -4 mother of runes, 2 swords to plowshares in: 2 Gideon, 2 councils Judgement, 1 leonin-relic, 1 pithing needle"
-    },
-    {
       "id": 3,
       "name": "Storm",
       "user_id": 3,
@@ -28,15 +22,39 @@
   ],
   "cards": [
     {
-      "id": 1,
-      "deckId": 3,
-      "card_name": "Assasins Trophy",
-      "quantity": 4
+      "id": 2,
+      "deckId": 1,
+      "card_name": "bolt",
+      "quantity": "2"
     },
     {
       "id": 2,
       "deckId": 1,
-      "card_name": "bolt",
+      "card_name": "helix",
+      "quantity": "2"
+    },
+    {
+      "id": 2,
+      "deckId": 1,
+      "card_name": "tundra",
+      "quantity": "2"
+    },
+    {
+      "id": 2,
+      "deckId": 1,
+      "card_name": "island",
+      "quantity": "2"
+    },
+    {
+      "id": 2,
+      "deckId": 1,
+      "card_name": "mountain",
+      "quantity": "2"
+    },
+    {
+      "id": 2,
+      "deckId": 1,
+      "card_name": "plains",
       "quantity": "2"
     }
   ]
